+++ golang/testsuite/main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"fmt"
-	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/execution_impl"
+	"github.com/kurtosis-tech/kurtosis-testsuite-starter-pack/golang/testsuite/execution_impl"
 	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/lib/execution"
 	"github.com/sirupsen/logrus"
 	"os"