+++ golang/testsuite/testsuite_impl/advanced_network_test/advanced_network_test_.go
@@ -2,7 +2,7 @@ package advanced_network_test
 
 import (
 	"github.com/kurtosis-tech/kurtosis-client/golang/lib/networks"
-	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/networks_impl"
+	"github.com/kurtosis-tech/kurtosis-testsuite-starter-pack/golang/testsuite/networks_impl"
 	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/lib/testsuite"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"