+++ docs/changelog.md
@@ -3,8 +3,12 @@
 * Upgraded to Kurtosis 1.18, which allows for multiple instances of Kurtosis to be run at the same time
 * Upgraded testsuite API lib 0.2.0 -> 0.4.0
 
-### Removed
+### Fixes
+* Correct all old references to `kurtosis-libs` -> `kurtosis-testsuite-starter-pack`
+
+### Removals
 * Removed Apache-2 license, dedicating everything inside this example to the public domain
+* Now-unused testsuite API `.proto` file
 
 # 1.32.1
 ### Changes